refactoring and implementation of 2 enhancements
